Add auto publish gem starter workflow
This is just an experiment for now. I'm trying to see if these starter
workflows can be a key part of making our GitHub actions more
discoverable, including details about how you a team might want to use
them.

I followed along with the GitHub guide[^1] for creating these "starter
workflows" that are private to our org.

The workflow can be copied via the GitHub web UI for any repo in our
org, and the included metadata helps GitHub suggest certain workflows
and organize them based on the repo trying to add a workflow.

[^1]: https://docs.github.com/en/actions/using-workflows/creating-starter-workflows-for-your-organization