Add update-action-tags workflow
GitHub Actions have this convention where they create major (e.g. `v1`)
and major.minor (e.g. `v1.2`) that point to the latest full semver
release (e.g. `v1.2.3`). Moving tags can be a bit of a pain, but there's
a nice action that can keep things updated for us.

This starter workflow is mostly a copy from something we have in
balto-utils but makes it a bit easier to adopt into a new Action repo.
It's worth noting that the readme for the action we're using does have
suggestions for creating and moving -alpha, -beta, and -rc tags, but
since we don't really use those, I left them off for now.